Fixed regression from migration to new parsers.  Most of the new parsers parsed the file size as an integer instead of a long.  Changed all of them to set the size to long.  This problem was detected by the reporter of:   http://issues.apache.org/bugzilla/show_bug.cgi?id=30345
Affected files:
  EnterpriseUnixFTPEntryParser.java
  NTFTPEntryParser.java
  OS2FTPEntryParser.java
  OS400FTPEntryParser.java
  UnixFTPEntryParser.java
  VMSFTPEntryParser.java

PR: 30345
Submitted by: kurt.kessel@cmediacorp.com (Kurt Kessel)


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/net/trunk@139520 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/net/ftp/parser/EnterpriseUnixFTPEntryParser.java
@@ -27,7 +27,7 @@ import org.apache.commons.net.ftp.FTPFile;
  * DefaultFTPParserFactory by classname.  It will not be chosen
  * by the autodetection scheme.
  * </B>
- * @version $Id: EnterpriseUnixFTPEntryParser.java,v 1.11 2004/04/21 23:30:33 scohen Exp $
+ * @version $Id: EnterpriseUnixFTPEntryParser.java,v 1.12 2004/07/28 05:01:47 dfs Exp $
  * @author <a href="Winston.Ojeda@qg.com">Winston Ojeda</a>
  * @see org.apache.commons.net.ftp.FTPFileEntryParser FTPFileEntryParser (for usage instructions)
  * @see org.apache.commons.net.ftp.parser.DefaultFTPFileEntryParserFactory
@@ -101,7 +101,7 @@ public class EnterpriseUnixFTPEntryParser extends RegexFTPFileEntryParserImpl
             file.setGroup(grp);
             try
             {
-                file.setSize(Integer.parseInt(filesize));
+                file.setSize(Long.parseLong(filesize));
             }
             catch (NumberFormatException e)
             {
